// jeffrey-microscope/pages-microscope/src/services/api/model/WorkspaceEventType.ts
/*
 * Jeffrey
 * Copyright (C) 2025 Petr Bouda
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

enum WorkspaceEventType {
  PROJECT_CREATED = 'PROJECT_CREATED',
  PROJECT_DELETED = 'PROJECT_DELETED',
  PROJECT_INSTANCE_CREATED = 'PROJECT_INSTANCE_CREATED',
  PROJECT_INSTANCE_SESSION_CREATED = 'PROJECT_INSTANCE_SESSION_CREATED',
  PROJECT_INSTANCE_SESSION_DELETED = 'PROJECT_INSTANCE_SESSION_DELETED',
  PROJECT_INSTANCE_SESSION_FINISHED = 'PROJECT_INSTANCE_SESSION_FINISHED',
  PROJECT_INSTANCE_SESSION_RECORDING_FILE_CREATED = 'PROJECT_INSTANCE_SESSION_RECORDING_FILE_CREATED',
  PROJECT_INSTANCE_SESSION_ARTIFACT_FILE_CREATED = 'PROJECT_INSTANCE_SESSION_ARTIFACT_FILE_CREATED',
  PROJECT_INSTANCE_FINISHED = 'PROJECT_INSTANCE_FINISHED',
  PROJECT_INSTANCE_EXPIRED = 'PROJECT_INSTANCE_EXPIRED'
}

export default WorkspaceEventType;